Change Log (2018-05-18): - Collision resolution kind of working (math is hard)

+++ 4or/main.cpp
@@ -69,7 +69,7 @@ int main(int argc, char *argv[]) {
 	while (!glfwWindowShouldClose(window)) {
 
 		///Calculate frame time
-		GLfloat currentFrame = glfwGetTime();
+		GLfloat currentFrame = (GLfloat)glfwGetTime();
 		deltaTime = currentFrame - lastFrame;
 		lastFrame = currentFrame;
 		glfwPollEvents();
